1.优化了选定文件发生变化时的卡顿现象. 2.递归浏览现在将作为右键菜单的功能之一被使用. 3.优化了ExtX解析速度.

+++ SingularityForensic/FAT/FatStreamParsingProvider.cs
@@ -514,7 +514,7 @@ namespace SingularityForensic.FAT {
             StFatClusterNode? lastCluster = null;
             long firstClusterLBA = 0;
 
-            int blockCount = 0;
+            long blockCount = 0;
             ulong lastClusterNum = 0;
 
             //重置局部变量;参数为本次(循环)最新的独立头块;
